Fixing log entries

+++ ras_common/ons_database.py
@@ -26,7 +26,7 @@ class ONSDatabase(object):
         self._base = declarative_base()
         self._session = scoped_session(sessionmaker())
 
-    def log(self, text):
+    def info(self, text):
         self._env.logger.info('[db] {}'.format(text))
 
     def warn(self, text):
@@ -34,13 +34,16 @@ class ONSDatabase(object):
 
     def activate(self):
         if self._env.get('enable_database', 'false').lower() not in ['true', 'yes']:
-            return self.log('Database is NOT enabled [missing "enabled_database = true"]')
+            return self.info('Database is NOT enabled [missing "enabled_database = true"]')
 
         if not self.check_paths():
             return self.warn('[swagger_server/models_local/_models.py] file is missing')
 
         db_connection = self._env.get('db_connection')
-        self.log('Database connection is "{}"'.format(db_connection))
+        if not db_connection:
+            return self.warn('Database connection not available [{}]'.format(db_connection))
+
+        self.info('Database connection is "{}"'.format(db_connection))
         self._engine = create_engine(db_connection, convert_unicode=True)
         self._session.remove()
         if self._env.drop_database:
@@ -55,28 +58,28 @@ class ONSDatabase(object):
         return Path('swagger_server/models_local/_models.py').is_file()
 
     def drop(self):
-        self.log('Dropping any existing Database Tables')
+        self.info('Dropping any existing Database Tables')
         from swagger_server.models_local import _models
         connection = self._env.get('db_connection')
         schema = self._env.get('db_schema')
         if connection.startswith('postgres'):
-            self.log('Dropping pre-existing schema "{}" if it exists'.format(schema))
+            self.info('Dropping pre-existing schema "{}" if it exists'.format(schema))
             self._base.metadata.schema = schema
             self._engine.execute("DROP SCHEMA IF EXISTS {} CASCADE".format(schema))
         else:
             self._base.metadata.drop_all(self._engine)
 
     def create(self):
-        self.log('Creating any missing Database tables')
+        self.info('Creating any missing Database tables')
         from swagger_server.models_local import _models
         connection = self._env.get('db_connection')
         schema = self._env.get('db_schema')
         if connection.startswith('postgres'):
-            self.log('Creating Database schema "{}"'.format(schema))
+            self.info('Creating Database schema "{}"'.format(schema))
             self._base.metadata.schema = schema
 
         if connection.startswith('postgres'):
-            self.log("Creating schema {} if it does't exist".format(schema))
+            self.info("Creating schema {} if it does't exist".format(schema))
             self._engine.execute("CREATE SCHEMA IF NOT EXISTS {}".format(schema))
-        self.log("Running Create-All")
+        self.info("Running Create-All")
         self._base.metadata.create_all(self._engine)
